Divide code into multiple files (#16)
* Put `clear_line()` in `utils.py`

* Create `chatgpt.py` & add related functions

* Add comments

* Reorder imports
