+++ chatgpt.py
@@ -1,0 +1,13 @@
+import openai
+
+
+def init(api_key):
+    openai.api_key = api_key
+
+
+def chat(prompt):
+    response = openai.ChatCompletion.create(
+        model="gpt-3.5-turbo",
+        messages=[{"role": "user", "content": prompt}]
+    )
+    return response.choices[0].message.content.strip()
